LayoutTests/fast/dom/html-attribute-types.html

<head>
<script src="../js/resources/js-test-pre.js"></script>
<script>
var img;
var pre;
var basefont;

function test() {
    if (window.testRunner)
        testRunner.dumpAsText();

    img = document.createElement('img');
    shouldBe("typeof img.border", "'string'");

    pre = document.createElement('pre');
    shouldBe("typeof pre.width", "'number'");
    
    basefont = document.createElement('basefont');
    shouldBe("typeof basefont.size", "'number'")
}
</script>
</head>
<body onload="test()">
This checks the types of some HTML DOM attributes whose types currently clash with our implementation.
<div id="console"></div>
</body>
